Move title migration to a HOC
