+++ assets/js/base/components/filter-update-heading/style.scss
@@ -2,7 +2,7 @@
 	background: transparent;
 	border-radius: 0;
 	border-width: 0 0 1px;
-	margin: 1px calc(-1px - $gap);
+	margin: 1px 0;
 
 	.block-editor-warning__actions {
 		margin-top: 0;
